feat(battle): achieve live pcap parity and refactor engine
- full battle1 + battle2 parity against captured live pcap (type=9, type=111, type=335, type=350 effect slices all match)
- split battle engine by concern:
  - mechanics/: bloodtithe, channel (incl. MonitorContinueChannel), injury_counter, nuodika, round_end, shadowcloak
  - manager/: stateful managers only (round/card/calculate/fight_data/ex_point/entity/buff)
  - steps/: stateless step utilities (broadcast, trigger_embed, ex_gain, step_normalize)
- round_mgr.rs reduced from ~3.7k to ~1.8k lines; zero hero-specific identifiers remain
- new config tables wired: buff_act, magic_circle, rule, skill_next, skill_bufftype, skill_behavior_condition, fight_const, fight_card_choice, monster_group, resistances_attribute, store_recommend
- protocol: string-or-number deserializer for int64/uint64 + variant-name deserializer for enum-tagged i32 fields
- battle_gen tool for replay generation and parity diffing
- database/sdkserver/handlers updates supporting the above

+++ database/src/db/game/currencies.rs
@@ -85,7 +85,7 @@ pub async fn add_currency(
     .bind(user_id)
     .bind(currency_id)
     .bind(amount)
-    .bind(timestamp as i64)
+    .bind(timestamp)
     .execute(pool)
     .await?;
     Ok(())
@@ -115,7 +115,7 @@ pub async fn remove_currency(
          WHERE user_id = ? AND currency_id = ?",
     )
     .bind(amount)
-    .bind(timestamp as i64)
+    .bind(timestamp)
     .bind(user_id)
     .bind(currency_id)
     .execute(pool)
@@ -142,7 +142,7 @@ pub async fn set_currency(
     .bind(user_id)
     .bind(currency_id)
     .bind(quantity)
-    .bind(timestamp as i64)
+    .bind(timestamp)
     .execute(pool)
     .await?;
     Ok(())
